Wrapping new workspaces view in feature flag
