+++ components/dashboard/src/app/AppRoutes.tsx
@@ -50,7 +50,8 @@ import { BlockedRepositories } from "../admin/BlockedRepositories";
 import PersonalAccessTokenCreateView from "../settings/PersonalAccessTokensCreateView";
 import { StartWorkspaceModalContext } from "../workspaces/start-workspace-modal-context";
 import { StartWorkspaceOptions } from "../start/start-workspace-options";
-import { WebsocketClients } from "./WebsocketSubscriptions";
+import { WebsocketClients } from "./WebsocketClients";
+import { useFeatureFlags } from "../contexts/FeatureFlagContext";
 
 const Setup = React.lazy(() => import(/* webpackPrefetch: true */ "../Setup"));
 const Workspaces = React.lazy(() => import(/* webpackPrefetch: true */ "../workspaces/Workspaces"));
@@ -100,6 +101,7 @@ export const AppRoutes: FunctionComponent<AppRoutesProps> = ({ user, teams }) =>
     const hash = getURLHash();
     const { startWorkspaceModalProps, setStartWorkspaceModalProps } = useContext(StartWorkspaceModalContext);
     const [isWhatsNewShown, setWhatsNewShown] = useState(shouldSeeWhatsNew(user));
+    const { useNewWorkspacesList } = useFeatureFlags();
 
     // Prefix with `/#referrer` will specify an IDE for workspace
     // We don't need to show IDE preference in this case
@@ -144,7 +146,6 @@ export const AppRoutes: FunctionComponent<AppRoutesProps> = ({ user, teams }) =>
                 />
             );
         } else {
-            // return <div>create workspace yay {hash}</div>;
             return <CreateWorkspace contextUrl={hash} />;
         }
     }
@@ -174,8 +175,11 @@ export const AppRoutes: FunctionComponent<AppRoutesProps> = ({ user, teams }) =>
                     <Route path={projectsPathNew} exact component={NewProject} />
                     <Route path="/open" exact component={Open} />
                     <Route path="/setup" exact component={Setup} />
-                    <Route path={workspacesPathMain} exact component={WorkspacesNew} />
-                    <Route path={workspacesPathMain} exact component={Workspaces} />
+                    <Route
+                        path={workspacesPathMain}
+                        exact
+                        component={useNewWorkspacesList ? WorkspacesNew : Workspaces}
+                    />
                     <Route path={settingsPathAccount} exact component={Account} />
                     <Route path={usagePathMain} exact component={Usage} />
                     <Route path={settingsPathIntegrations} exact component={Integrations} />
@@ -268,8 +272,9 @@ export const AppRoutes: FunctionComponent<AppRoutesProps> = ({ user, teams }) =>
                                     switch (maybeProject) {
                                         case "projects":
                                             return <Projects />;
+                                        // TODO: Determine if we surface this - if so, can it also use WorkspacesNew?
                                         case "workspaces":
-                                            return <Workspaces />;
+                                            return useNewWorkspacesList ? <WorkspacesNew /> : <Workspaces />;
                                         case "members":
                                             return <Members />;
                                         case "settings":
